feature-2

+++ app/src/main/java/com/myapp/alarm/data/local/AlarmDao.kt
@@ -25,6 +25,6 @@ interface AlarmDao {
         suspend fun  updateAlarm ( alarm: Alarm)
 
        @Query("Update Alarm set isScheduled = :bool where id = :id")
-         suspend fun updateSchedule(bool : Boolean, id : Int)
+         suspend fun updateSchedule(bool : Boolean, id : Long)
 
 }
